% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobMReg-package.r
\docType{package}
\name{RobMReg-package}
\alias{RobMReg}
\alias{RobMReg-package}
\title{RobMReg: TODO:title}
\description{
To learn more about simlab, start with the vignettes:
\code{browseVignettes(package = "RobMReg")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.unimib.it}
  \item \url{https://github.com/giannamonti/RobMReg}
  \item Report bugs at \url{https://github.com/giannamonti/RobMReg/issues}
}

}
\author{
\strong{Maintainer}: Gianna Monti \email{gianna.monti@unimib.it}

}
\keyword{internal}
