% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmethodlassoEfast.R
\name{runmethodlassoEfast}
\alias{runmethodlassoEfast}
\title{run method lasso Efast}
\usage{
runmethodlassoEfast(
  Y,
  X,
  typeset = "square",
  trimset = 1,
  M = 50,
  q = 0.2,
  filter = FALSE
)
}
\arguments{
\item{Y}{.}

\item{X}{.}

\item{typeset}{.}

\item{trimset}{.}

\item{M}{.}

\item{q}{.}

\item{filter}{.}
}
\value{
result
}
\description{
run method lasso Efast
}
\examples{
NULL
}
